mutants/pyproject.toml
[project]
name = "home-finder"
version = "0.1.0"
description = "Multi-platform London rental property scraper with commute filtering and Telegram notifications"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "crawlee[beautifulsoup]>=1.3.2b4",
    "traveltimepy>=4.0",
    "aiogram>=3.4",
    "pydantic>=2.6",
    "pydantic-settings>=2.2",
    "aiosqlite>=0.20",
    "structlog>=24.1",
    "curl-cffi>=0.7",
    "httpx>=0.27",
    "anthropic>=0.77.0",
    "imagehash>=4.3.1",
    "Pillow>=10.0.0",
    "fastapi>=0.115",
    "uvicorn[standard]>=0.34",
    "jinja2>=3.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "pytest-httpx>=0.35",
    "pytest-cov>=6.0",
    "hypothesis>=6.100",
    "pytest-playwright>=0.6",
    "pytest-socket>=0.7",
    "pytest-xdist>=3.5",
    "mutmut>=3.0",
    "ruff>=0.8",
    "mypy>=1.13",
]

[project.scripts]
home-finder = "home_finder.main:main"

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
testpaths = ["tests"]
addopts = "--tb=short -p no:cacheprovider -m 'not slow and not browser' --disable-socket --allow-unix-socket"
markers = [
    "unit: Unit tests (fast, no I/O)",
    "integration: Integration tests (may use in-memory DB)",
    "slow: Slow tests (network, real scraping)",
    "e2e: End-to-end tests combining multiple pipeline stages",
    "browser: Browser-based E2E tests (requires Playwright)",
    "allow_hosts: Allow specific hosts for network access (pytest-socket)",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]

[tool.ruff.lint.per-file-ignores]
"src/home_finder/web/routes.py" = ["B008"]  # FastAPI Depends() in defaults

[tool.mypy]
python_version = "3.11"
strict = true

[tool.coverage.run]
source = ["src/home_finder"]
branch = true
omit = [
    "src/home_finder/logging.py",
    "src/home_finder/data/*",
]

[tool.coverage.report]
fail_under = 75
show_missing = true
exclude_lines = [
    "pragma: no cover",
    "if TYPE_CHECKING:",
    "if __name__ ==",
]

[tool.mutmut]
paths_to_mutate = [
    "src/home_finder/filters/criteria.py",
]
also_copy = ["src/"]
pytest_add_cli_args = ["-o", "addopts=--tb=short -p no:cacheprovider -m 'not slow and not browser' --disable-socket --allow-unix-socket --ignore=tests/integration"]
debug = true

[tool.uv]
package = true

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/home_finder"]
